refactor `dsc` main to be multiple files
